+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Updates
 - Use `MSBuildTasks` by default and when fail use `DotNetTasks` as fallback. (Rollback: #85)
 - Add `Tools.MSBuild` to support MSBuild from Visual Studio 2026. (Fix: #90)
-- Add `BuildExtension.UseDotnetToBuild` to prefer to use `dotnet build` instead of `msbuild`.
+- Add `BuildExtension.BuildTools` with list of tools to build in order by `msbuild` and `dotnet`.
 
 ## [1.11.3] / 2026-04-16
 ### Updated